beta-version of application

+++ src/main/java/com/simbirsoft/intership/service/TaskServiceImpl.java
@@ -3,15 +3,21 @@ package com.simbirsoft.intership.service;
 import com.simbirsoft.intership.dto.request.ChangeTaskStatusDto;
 import com.simbirsoft.intership.dto.request.CreatingTaskDto;
 import com.simbirsoft.intership.dto.request.UpdatingTaskDto;
+import com.simbirsoft.intership.dto.response.TaskDto;
+import com.simbirsoft.intership.exception.NotFoundPermissionException;
+import com.simbirsoft.intership.exception.NotFoundTaskException;
+import com.simbirsoft.intership.exception.ProjectClosedException;
 import com.simbirsoft.intership.model.ProjectPermission;
 import com.simbirsoft.intership.model.Realise;
 import com.simbirsoft.intership.model.Task;
 import com.simbirsoft.intership.model.User;
 import com.simbirsoft.intership.model.enumaration.Permission;
+import com.simbirsoft.intership.model.enumaration.ProjectStatus;
 import com.simbirsoft.intership.model.enumaration.TaskStatus;
 import com.simbirsoft.intership.repository.TaskRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -20,15 +26,19 @@ public class TaskServiceImpl implements TaskService {
     private final RealiseService realiseService;
     private final TaskRepository taskRepository;
 
+    @Transactional
     @Override
-    public Task createTask(CreatingTaskDto creatingTaskDto, User user) {
+    public TaskDto createTask(CreatingTaskDto creatingTaskDto, User user) {
         ProjectPermission permissionLeading = projectPermissionService
                 .getPermission(user.getId(), creatingTaskDto.getProjectId(), Permission.Leading);
+        if (permissionLeading.getProject().getProjectStatus() == ProjectStatus.CLOSED) {
+            throw new ProjectClosedException(creatingTaskDto.getProjectId(), user.getId());
+        }
         ProjectPermission permissionPerformer = projectPermissionService
                 .getPermission(user.getId(), creatingTaskDto.getProjectId());
         Realise realise = realiseService.getRealise(creatingTaskDto.getRealiseId(),
                 permissionLeading.getProject().getId());
-        return taskRepository.save(Task.builder()
+        Task task = taskRepository.save(Task.builder()
                 .owner(user)
                 .taskStatus(TaskStatus.IN_PROGRESS)
                 .title(creatingTaskDto.getTitle())
@@ -37,50 +47,61 @@ public class TaskServiceImpl implements TaskService {
                 .project(permissionLeading.getProject())
                 .realise(realise)
                 .build());
+        return TaskDto.from(task, realise, permissionLeading.getProject());
     }
 
+    @Transactional
     @Override
-    public Task updateTask(long taskId, UpdatingTaskDto updatingTaskDto, User user) {
-        ProjectPermission projectPermission = projectPermissionService.getPermission(
+    public TaskDto updateTask(long taskId, UpdatingTaskDto updatingTaskDto, User user) {
+        ProjectPermission permissionOwner = projectPermissionService.getPermission(
                 user.getId(),
                 updatingTaskDto.getProjectId()
         );
+        if (permissionOwner.getProject().getProjectStatus() == ProjectStatus.CLOSED) {
+            throw new ProjectClosedException(updatingTaskDto.getProjectId(), user.getId());
+        }
 
-        Task task = taskRepository.findTaskByIdAndProjectId(taskId, projectPermission.getProject().getId())
-                .orElseThrow();//todo exception
+        Task task = taskRepository.findTaskByIdAndProjectId(taskId, permissionOwner.getProject().getId())
+                .orElseThrow(() -> new NotFoundTaskException(taskId, updatingTaskDto.getProjectId()));
         ProjectPermission permissionPerformer = projectPermissionService
                 .getPermission(updatingTaskDto.getPerformerId(), updatingTaskDto.getProjectId());
-
-        task.setRealise(realiseService.getRealise(updatingTaskDto.getRealiseId(),
-                projectPermission.getProject().getId()));
+        Realise realise = realiseService.getRealise(updatingTaskDto.getRealiseId(),
+                permissionOwner.getProject().getId());
+        task.setRealise(realise);
         task.setPerformer(permissionPerformer.getUser());
         task.setTitle(updatingTaskDto.getTitle());
 
-        return taskRepository.save(task);
+        return TaskDto.from(taskRepository.save(task), realise, permissionOwner.getProject());
     }
 
+    @Transactional
     @Override
-    public Task changeTaskStatus(long taskId, ChangeTaskStatusDto changeTaskStatusDto, User user) {
+    public void changeTaskStatus(long taskId, ChangeTaskStatusDto changeTaskStatusDto, User user) {
         ProjectPermission projectPermission = projectPermissionService.getPermission(
                 user.getId(),
                 changeTaskStatusDto.getProjectId()
         );
+        if (projectPermission.getProject().getProjectStatus() == ProjectStatus.CLOSED) {
+            throw new ProjectClosedException(changeTaskStatusDto.getProjectId(), user.getId());
+        }
         Task task = taskRepository.findTaskByIdAndProjectId(taskId, projectPermission.getProject().getId())
-                .orElseThrow();//todo exception
+                .orElseThrow(() -> new NotFoundTaskException(taskId, changeTaskStatusDto.getProjectId()));
 
-        checkAccessChangeStatusTask(projectPermission, task);
+        hasAccessChangeStatusTask(projectPermission, task);
 
         task.setTaskStatus(changeTaskStatusDto.getTaskStatus());
-        return taskRepository.save(task);
+        taskRepository.save(task);
     }
 
-    private void checkAccessChangeStatusTask(ProjectPermission projectPermission, Task task) {
+    private void hasAccessChangeStatusTask(ProjectPermission projectPermission, Task task) {
+        //статус может менять пользователь Leading, испольняющий и владелец проекта
         if (!projectPermission.getPermissions().contains(Permission.Leading) &&
                 !task.getPerformer().getId().equals(projectPermission.getUser().getId())) {
-            throw new IllegalArgumentException(); //todo exception
+            throw new NotFoundPermissionException(task.getPerformer().getId(), projectPermission.getProject().getId());
         }
     }
 
+    @Transactional
     @Override
     public void deleteTask(long taskId, long projectId, User user) {
         ProjectPermission projectPermission = projectPermissionService.getPermission(
@@ -88,8 +109,11 @@ public class TaskServiceImpl implements TaskService {
                 projectId,
                 Permission.Leading
         );
-        //todo exception
-        Task task = taskRepository.findTaskByIdAndProjectId(taskId, projectId).orElseThrow();
+        if (projectPermission.getProject().getProjectStatus() == ProjectStatus.CLOSED) {
+            throw new ProjectClosedException(projectId, user.getId());
+        }
+        Task task = taskRepository.findTaskByIdAndProjectId(taskId, projectId)
+                .orElseThrow(() -> new NotFoundTaskException(taskId, projectId));
         taskRepository.delete(task);
     }
 }
